lambda/processor/requirements.txt
slack_sdk>=3.27.0,<4.0.0
strands-agents>=0.1.0
strands-agents-builder>=0.1.0
awslabs.billing-cost-management-mcp-server
